+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Handle closed channel in rabbitMQ manager
 - Move Record status update to standard batch query and prevent stale data
 - Handle `None` values in array of author names
+- Update DOAB ingest process to consume OAI_DC records
 
 ## 2021-04-09 -- v0.5.4
 ### Fixed